Chris Ali
- Cleaned up EngineModel and FixedPitchPropEngine classes to include one
method to update the engine's members, and getter methods for each
member
- Added derivative states, and engine parameters to the simOut
(originally outputState) output double[]; made this a class member of
Integrate6DOFEquations and added a getter method. This will be used to
transmit to other threads or across a network connection  
- Reworked the plotting methods by creating a PlotUtilities class to
create a HashMap of XYPlots. This can be accessed by SimulationPlots to
generate a group of subplots on one window. Future updates will allow
for multiple plot windows to be generated as desired.