Chris Ali
- Fixed Keyboard and Mouse classes, which were not polling correctly.
The cause was from using raw input instead of DirectInput. Raw input
works if a Java Window is available to be selected, but if not, then the
controller will never get polled. To use DirectInput, ensure the
following VM arguements are added:

	-Djava.library.path=./dist/
	-Djinput.useDefaultPlugin=false 
	-Djinput.plugins=net.java.games.input.DirectInputEnvironmentPlugin

- Added mouse and keyboard inputs into simulation. The mouse can control
elevator, aileron, and throttle, and the keyboard can control all those
plus pause and reset. These controllers still need to be tested more in
depth.

- Integrated pause and reset into the simulation. The button sensitivity
is difficult to work with due to the fast simulation step time, so a
better method of toggling control presses is needed